Fix mixed case on Ory+Istio not creating AP and RA for non-jwt handlers
