+++ internal/processing/istio/istio_suite_test.go
@@ -1,0 +1,15 @@
+package istio
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
+)
+
+func TestOry(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecsWithDefaultAndCustomReporters(t, "Istio Suite",
+		[]Reporter{printer.NewProwReporter("api-gateway-istio-testsuite")})
+}
