+++ internal/processing/istio/reconciliation.go
@@ -5,7 +5,6 @@ import (
 
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	"github.com/kyma-incubator/api-gateway/internal/processing"
-	"github.com/kyma-incubator/api-gateway/internal/processing/processors"
 	"github.com/kyma-incubator/api-gateway/internal/validation"
 	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -17,10 +16,10 @@ type Reconciliation struct {
 }
 
 func NewIstioReconciliation(config processing.ReconciliationConfig) Reconciliation {
-	vsProcessor := NewVirtualServiceProcessor(config)
-	apProcessor := processors.NewAuthorizationPolicyProcessor(config)
-	raProcessor := processors.NewRequestAuthenticationProcessor(config)
 	acProcessor := NewAccessRuleProcessor(config)
+	vsProcessor := NewVirtualServiceProcessor(config)
+	apProcessor := NewAuthorizationPolicyProcessor(config)
+	raProcessor := NewRequestAuthenticationProcessor(config)
 
 	return Reconciliation{
 		processors: []processing.ReconciliationProcessor{vsProcessor, raProcessor, apProcessor, acProcessor},
