+++ internal/processing/processors/authorization_policy_processor.go
@@ -14,71 +14,13 @@ import (
 
 // AuthorizationPolicyProcessor is the generic processor that handles the Istio Authorization Policies in the reconciliation of API Rule.
 type AuthorizationPolicyProcessor struct {
-	Creator authorizationPolicyCreator
+	Creator AuthorizationPolicyCreator
 }
 
-// NewAuthorizationPolicyProcessor returns a AuthorizationPolicyProcessor with the desired state handling specific for the Istio handler.
-func NewAuthorizationPolicyProcessor(config processing.ReconciliationConfig) AuthorizationPolicyProcessor {
-	return AuthorizationPolicyProcessor{
-		Creator: authorizationPolicyCreator{
-			additionalLabels: config.AdditionalLabels,
-		},
-	}
-}
-
-type authorizationPolicyCreator struct {
-	additionalLabels map[string]string
-}
-
-// Create returns the Authorization Policy using the configuration of the APIRule.
-func (r authorizationPolicyCreator) Create(api *gatewayv1beta1.APIRule) map[string]*securityv1beta1.AuthorizationPolicy {
-	pathDuplicates := HasPathDuplicates(api.Spec.Rules)
-	authorizationPolicies := make(map[string]*securityv1beta1.AuthorizationPolicy)
-	for _, rule := range api.Spec.Rules {
-		if processing.IsSecured(rule) {
-			ar := generateAuthorizationPolicy(api, rule, r.additionalLabels)
-			authorizationPolicies[getAuthorizationPolicyKey(pathDuplicates, ar)] = ar
-		}
-	}
-	return authorizationPolicies
-}
-
-func generateAuthorizationPolicy(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, additionalLabels map[string]string) *securityv1beta1.AuthorizationPolicy {
-	namePrefix := fmt.Sprintf("%s-", api.ObjectMeta.Name)
-	namespace := api.ObjectMeta.Namespace
-	ownerRef := processing.GenerateOwnerRef(api)
-
-	apBuilder := builders.AuthorizationPolicyBuilder().
-		GenerateName(namePrefix).
-		Namespace(namespace).
-		Owner(builders.OwnerReference().From(&ownerRef)).
-		Spec(builders.AuthorizationPolicySpecBuilder().From(generateAuthorizationPolicySpec(api, rule))).
-		Label(processing.OwnerLabel, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace)).
-		Label(processing.OwnerLabelv1alpha1, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace))
-
-	for k, v := range additionalLabels {
-		apBuilder.Label(k, v)
-	}
-
-	return apBuilder.Get()
-}
-
-func generateAuthorizationPolicySpec(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule) *v1beta1.AuthorizationPolicy {
-	var serviceName string
-	if rule.Service != nil {
-		serviceName = *rule.Service.Name
-	} else {
-		serviceName = *api.Spec.Service.Name
-	}
-
-	authorizationPolicySpec := builders.AuthorizationPolicySpecBuilder().
-		Selector(builders.SelectorBuilder().MatchLabels("app", serviceName)).
-		Rule(builders.RuleBuilder().
-			RuleFrom(builders.RuleFromBuilder().Source()).
-			RuleTo(builders.RuleToBuilder().
-				Operation(builders.OperationBuilder().Methods(rule.Methods).Path(rule.Path))))
-
-	return authorizationPolicySpec.Get()
+// AuthorizationPolicyCreator provides the creation of AuthorizationPolicies using the configuration in the given APIRule.
+// The key of the map is expected to be unique and comparable with the
+type AuthorizationPolicyCreator interface {
+	Create(api *gatewayv1beta1.APIRule) map[string]*securityv1beta1.AuthorizationPolicy
 }
 
 func (r AuthorizationPolicyProcessor) EvaluateReconciliation(ctx context.Context, client ctrlclient.Client, apiRule *gatewayv1beta1.APIRule) ([]*processing.ObjectChange, error) {
@@ -109,7 +51,7 @@ func (r AuthorizationPolicyProcessor) getActualState(ctx context.Context, client
 	pathDuplicates := HasPathDuplicates(api.Spec.Rules)
 	for i := range apList.Items {
 		obj := apList.Items[i]
-		authorizationPolicies[getAuthorizationPolicyKey(pathDuplicates, obj)] = obj
+		authorizationPolicies[GetAuthorizationPolicyKey(pathDuplicates, obj)] = obj
 	}
 
 	return authorizationPolicies, nil
@@ -144,24 +86,55 @@ func (r AuthorizationPolicyProcessor) getObjectChanges(desiredAps map[string]*se
 	return apChangesToApply
 }
 
-func getAuthorizationPolicyKey(hasPathDuplicates bool, ap *securityv1beta1.AuthorizationPolicy) string {
+func GenerateAuthorizationPolicy(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, additionalLabels map[string]string) *securityv1beta1.AuthorizationPolicy {
+	namePrefix := fmt.Sprintf("%s-", api.ObjectMeta.Name)
+	namespace := api.ObjectMeta.Namespace
+	ownerRef := processing.GenerateOwnerRef(api)
+
+	apBuilder := builders.AuthorizationPolicyBuilder().
+		GenerateName(namePrefix).
+		Namespace(namespace).
+		Owner(builders.OwnerReference().From(&ownerRef)).
+		Spec(builders.AuthorizationPolicySpecBuilder().From(GenerateAuthorizationPolicySpec(api, rule))).
+		Label(processing.OwnerLabel, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace)).
+		Label(processing.OwnerLabelv1alpha1, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace))
+
+	for k, v := range additionalLabels {
+		apBuilder.Label(k, v)
+	}
+
+	return apBuilder.Get()
+}
+
+func GenerateAuthorizationPolicySpec(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule) *v1beta1.AuthorizationPolicy {
+	var serviceName string
+	if rule.Service != nil {
+		serviceName = *rule.Service.Name
+	} else {
+		serviceName = *api.Spec.Service.Name
+	}
+
+	authorizationPolicySpec := builders.AuthorizationPolicySpecBuilder().
+		Selector(builders.SelectorBuilder().MatchLabels("app", serviceName)).
+		Rule(builders.RuleBuilder().
+			RuleFrom(builders.RuleFromBuilder().Source()).
+			RuleTo(builders.RuleToBuilder().
+				Operation(builders.OperationBuilder().Methods(rule.Methods).Path(rule.Path))))
+
+	return authorizationPolicySpec.Get()
+}
+
+func GetAuthorizationPolicyKey(hasPathDuplicates bool, ap *securityv1beta1.AuthorizationPolicy) string {
 	key := ""
 	if ap.Spec.Rules != nil && len(ap.Spec.Rules) > 0 && ap.Spec.Rules[0].To != nil && len(ap.Spec.Rules[0].To) > 0 {
 		if hasPathDuplicates {
 			key = fmt.Sprintf("%s:%s",
-				sliceToString(ap.Spec.Rules[0].To[0].Operation.Paths),
-				sliceToString(ap.Spec.Rules[0].To[0].Operation.Methods))
+				processing.SliceToString(ap.Spec.Rules[0].To[0].Operation.Paths),
+				processing.SliceToString(ap.Spec.Rules[0].To[0].Operation.Methods))
 		} else {
-			key = sliceToString(ap.Spec.Rules[0].To[0].Operation.Paths)
+			key = processing.SliceToString(ap.Spec.Rules[0].To[0].Operation.Paths)
 		}
 	}
 
 	return key
-}
-
-func sliceToString(ss []string) (s string) {
-	for _, el := range ss {
-		s += el
-	}
-	return
 }
